requirements.txt
fastapi==0.116.1
pytz
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
Jinja2==3.1.6
PyJWT
SQLAlchemy
bcrypt
uvicorn==0.35.0
psycopg2==2.9.10
pydantic==2.11.7
annotated-types == 0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email-validator==2.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.1
setuptools==65.5.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.3
typer==0.16.1
typing_extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1